refactor: move editor, ncm, ast components to separate directories
