+++ src/components/ast/graph/TreeGraph.vue
@@ -1,9 +1,9 @@
 <script lang="ts">
 import { defineComponent } from "vue"
 import type { PropType } from "vue"
-import TreeGraphNode from "@/components/tree-graph/TreeGraphNode.vue"
-import { drawLinesFromNodeToChildrenOnSvgRecursively } from "@/components/tree-graph/domGraphicUtils"
-import type { TreeNode } from "@/components/tree-graph/types"
+import TreeGraphNode from "./TreeGraphNode.vue"
+import { drawLinesFromNodeToChildrenOnSvgRecursively } from "./domGraphicUtils"
+import type { TreeNode } from "./types"
 
 export default defineComponent({
   name: "TreeGraph",
